enable react-hot-reload for Explore, Dashboard, Profile and Sqllab
